fix(npe-portal): code-review fixes for cert detail/expiring/CSR features
Addresses 8 findings from the review of roadmap #2/#3/#4:

1. CRITICAL: list_filtered passed an i64 (bigint) into make_interval(days => $4),
   which Postgres can't resolve (no implicit bigint->integer cast) — the entire
   expiring drill-down 500'd at runtime. Cast to $4::int. Verified against the
   live DB (bigint errors, int works).
2. In-browser CSR generator silently dropped IPv6 and UPN/otherName SANs. Added
   IPv6 parsing and now throws a clear error listing any SAN it can't encode,
   instead of issuing a CSR missing a requested name. IPv6 round-trip verified.
3. Documented the PKCS#7/chain download's single-tier limitation honestly (the
   CA holds only its own cert) + POAM; both download paths use one source.
4. Corrected pkcs7.rs doc: certificates is a CMS SET OF, so order isn't
   positional (the prior "leaf first" promise was false).
5. Updated docs/compliance (RFC_COMPLIANCE, NIST_800-53_MAPPING) for the new
   PKCS#7 endpoint + shared ostrich_x509::pkcs7 module (CLAUDE.md mandate).
6. Extracted shared commonName() (lib/dn.ts, now handles escaped commas in DNs)
   and safeFileName() (lib/download.ts), removing 3 copies + a duplicated regex.
7. certificate-detail caInfo query is now lazy (enabled:false, fetched on the
   chain download) instead of firing on every detail view.
8. Added server-computed daysRemaining to the certificate list DTO/row; the
   expiring badge uses it so the list and detail view agree (no client recompute).

Also dropped redundant cert! assertions. cargo check + clippy clean; pkcs7 tests
pass; tsc + vite build clean (main bundle unchanged, pkijs stays lazy).

+++ crates/ostrich-db/src/repository/certificate.rs
@@ -310,7 +310,11 @@ impl CertificateRepository {
                     NOT revoked
                     AND not_before <= NOW()
                     AND not_after >= NOW()
-                    AND not_after < NOW() + make_interval(days => $4)
+                    -- $4 is bound as bigint; make_interval's `days` is integer and
+                    -- Postgres has no implicit bigint->integer cast in function
+                    -- resolution, so the ::int cast is required (not just on the
+                    -- NULL guard above).
+                    AND not_after < NOW() + make_interval(days => $4::int)
                 )
             )
         "#;
